requirements.txt
absl-py==1.4.0
aioice==0.7.6
aiortc==1.4.0
altair==4.2.2
attrs==22.2.0
av==10.0.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
dlib-binary
dnspython==2.3.0
entrypoints==0.4
flatbuffers==23.1.21
fonttools==4.38.0
gcloud==0.18.3
gitdb==4.0.10
GitPython==3.1.30
google-crc32c==1.5.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
imutils==0.5.4
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
mediapipe==0.9.1.0
netifaces==0.11.0
numpy==1.24.1
oauth2client==4.1.3
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydeck==0.8.0
pydub==0.25.1
pyee==9.0.4
Pygments==2.14.0
pylibsrtp==0.8.0
pymongo==4.3.3
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
Pyrebase4==4.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.1
rsa==4.9
scipy==1.10.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.17.0
streamlit-nested-layout==0.1.1
streamlit-option-menu==0.3.2
streamlit-webrtc==0.44.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wincertstore==0.2
wrapt==1.14.1
zipp==3.12.0
